requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.2.7            # via django
certifi==2020.4.5.2       # via requests
chardet==3.0.4            # via requests
dj-database-url==0.5.0    # via -r requirements.in
django-constance[redis]==2.6.0  # via -r requirements.in
django-crispy-forms==1.9.1  # via -r requirements.in
django-debug-toolbar==2.2  # via -r requirements.in
django-filer==1.7.1       # via -r requirements.in
django-js-asset==1.2.2    # via django-mptt
django-mptt==0.11.0       # via django-filer
django-polymorphic==2.1.2  # via django-filer
django-recaptcha==2.0.6   # via -r requirements.in
django-redis==4.12.1      # via -r requirements.in
django-smtp-ssl==1.0      # via -r requirements.in
django==3.0.7             # via -r requirements.in, django-debug-toolbar, django-filer, django-mptt, django-polymorphic, django-recaptcha, django-redis, djangorestframework, easy-thumbnails
djangorestframework==3.11.0  # via -r requirements.in
easy-thumbnails==2.7      # via django-filer
httpie==2.1.0             # via -r requirements.in
idna==2.9                 # via requests
mysqlclient==1.4.6        # via -r requirements.in
pillow==7.1.2             # via -r requirements.in, easy-thumbnails
pygments==2.6.1           # via httpie
python-dotenv==0.13.0     # via -r requirements.in
pytz==2020.1              # via django
redis==3.5.3              # via django-constance, django-redis
requests==2.23.0          # via httpie
six==1.15.0               # via django-filer
sqlparse==0.3.1           # via django, django-debug-toolbar
unidecode==1.1.1          # via django-filer
urllib3==1.25.9           # via requests
